TripKit (the core library) has no more external dependencies.
* ASPolygonKit code was added (and made internal)
* All RxSwift code has been moved to TripKitUI
* TKGeocoding + TKAutocompleting have been refactored to use completion blocks

The goal is to make it easier to use TripKit in smaller utilities without having to deal with dependencies. At SkedGo we use Rx more for UI-related things, though a couple of nice helpers are now in TripKitUI instead. By relying less on Rx, it should make a transition to Combine or the [upcoming async/await](https://forums.swift.org/t/swift-concurrency-roadmap/41611) a bit easier.